+++ Content.Server/_CE/ZLevels/Core/CEZLevelsSystem.API.cs
@@ -3,47 +3,16 @@
  * https://github.com/space-wizards/space-station-14/blob/master/LICENSE.TXT
  */
 
-using Content.Server._CE.PVS;
+using System.Linq;
+using Content.Server._CE.ZLevels.PVS;
 using Content.Shared._CE.ZLevels.Core.Components;
 using JetBrains.Annotations;
-using Robust.Shared.Map.Components;
 using Robust.Shared.Prototypes;
 
 namespace Content.Server._CE.ZLevels.Core;
 
 public sealed partial class CEZLevelsSystem
 {
-    /// <summary>
-    /// Initializes all uninitialized maps in the z-network.
-    /// </summary>
-    [PublicAPI]
-    public void InitializeZNetwork(Entity<CEZLevelsNetworkComponent> network)
-    {
-        foreach (var (_, mapUid) in network.Comp.ZLevels)
-        {
-            if (!TryComp<MapComponent>(mapUid, out var mapComp))
-            {
-                Log.Error($"Map entity {mapUid} does not have MapComponent.");
-                continue;
-            }
-
-            if (!_map.MapExists(mapComp.MapId))
-            {
-                Log.Error($"Map with ID {mapComp.MapId} does not exist.");
-                continue;
-            }
-
-            if (_map.IsInitialized(mapComp.MapId))
-            {
-                Log.Debug($"Map with ID {mapComp.MapId} is already initialized.");
-                continue;
-            }
-
-            _map.InitializeMap(mapComp.MapId);
-            Log.Info($"Map with ID {mapComp.MapId} initialized.");
-        }
-    }
-
     /// <summary>
     /// Creates a new entity zLevelNetwork
     /// </summary>
@@ -65,15 +34,15 @@ public sealed partial class CEZLevelsSystem
     /// </summary>
     private bool TryAddMapIntoZNetwork(Entity<CEZLevelsNetworkComponent> network, EntityUid mapUid, int depth)
     {
-        if (network.Comp.ZLevels.ContainsKey(depth))
-        {
-            Log.Error($"Failed to add map {mapUid} to ZLevelNetwork {network}: This depth is already occupied.");
-            return false;
-        }
-
         if (TryGetZNetwork(mapUid, out var otherNetwork))
         {
             Log.Error($"Failed attempt to add map {mapUid} to ZLevelNetwork {network}: This map is already in another network {otherNetwork}.");
+            return false;
+        }
+
+        if (network.Comp.ZLevels.ContainsKey(depth))
+        {
+            Log.Error($"Failed to add map {mapUid} to ZLevelNetwork {network}: This depth is already occupied.");
             return false;
         }
 
@@ -83,8 +52,13 @@ public sealed partial class CEZLevelsSystem
             return false;
         }
 
-        network.Comp.ZLevels.Add(depth, mapUid);
+        network.Comp.ZLevels[depth] = mapUid;
+        network.Comp.ZLevelByEntity[mapUid] = depth;
+
         Dirty(network);
+
+        var ev = new CEMapAddedIntoZNetworkEvent(network, depth);
+        RaiseLocalEvent(mapUid, ref ev);
 
         // Welcome to fast api code
         QuickApiCache(network, mapUid, depth);
@@ -101,9 +75,6 @@ public sealed partial class CEZLevelsSystem
 
         Dirty(mapUid, levelMapComponent);
 
-        var ev = new CEMapAddedIntoZNetworkEvent(network, depth);
-        RaiseLocalEvent(mapUid, ref ev);
-
         return true;
     }
 
@@ -116,8 +87,7 @@ public sealed partial class CEZLevelsSystem
                 success = false;
         }
 
-        var ev = new CEZLevelNetworkUpdatedEvent();
-        RaiseLocalEvent(network, ref ev);
+        RaiseLocalEvent(network, new CEZLevelNetworkUpdatedEvent());
 
         return success;
     }
@@ -159,13 +129,71 @@ public sealed partial class CEZLevelsSystem
 
         QueueDel(networkUid);
     }
+
+    private void QuickApiCache(Entity<CEZLevelsNetworkComponent> network, EntityUid value, int depth)
+    {
+        var comp = network.Comp;
+        var list = comp.SortedZLevels;
+
+        // Zero handling
+        if (comp.SortedMin == depth && comp.SortedMax == depth)
+        {
+            list.Add(value);
+            return;
+        }
+
+        var min = comp.SortedMin;
+        var max = comp.SortedMax;
+
+        if (depth < min)
+        {
+            var delta = min - depth;
+            if (delta == 1)
+            {
+                list.Insert(0, value);
+
+                comp.SortedMin = depth;
+                Dirty(network);
+                return;
+            }
+
+            list.InsertRange(0, Enumerable.Repeat(EntityUid.Invalid, delta - 1));
+            list.Insert(0, value);
+
+            comp.SortedMin = depth;
+            Dirty(network);
+            return;
+        }
+
+        if (depth > max)
+        {
+            var delta = depth - max;
+            if (delta == 1)
+            {
+                list.Add(value);
+
+                comp.SortedMax = depth;
+                Dirty(network);
+                return;
+            }
+
+            list.AddRange(Enumerable.Repeat(EntityUid.Invalid, delta - 1));
+            list.Add(value);
+
+            comp.SortedMax = depth;
+            Dirty(network);
+            return;
+        }
+
+        list[depth - min] = value;
+    }
 }
 
 /// <summary>
 /// Called on ZLevel Network Entity, when maps added or removed from network
 /// </summary>
-[ByRefEvent]
-public readonly struct CEZLevelNetworkUpdatedEvent;
+public sealed class CEZLevelNetworkUpdatedEvent : EntityEventArgs;
+
 
 /// <summary>
 /// Called on map, when it added to ZNetwork
