+++ Content.Server/_CE/ZLevels/Core/CEZLevelsSystem.View.cs
@@ -19,7 +19,7 @@ namespace Content.Server._CE.ZLevels.Core;
 public sealed partial class CEZLevelsSystem
 {
     [Dependency] private readonly IGameTiming _timing = default!;
-    //[Dependency] private readonly SharedActionsSystem _actions = default!;
+    [Dependency] private readonly SharedActionsSystem _actions = default!;
     [Dependency] private readonly ViewSubscriberSystem _viewSubscriber = default!;
     [Dependency] private readonly SharedPopupSystem _popup = default!;
 
@@ -58,13 +58,13 @@ public sealed partial class CEZLevelsSystem
 
     private void OnViewerInit(Entity<CEZLevelViewerComponent> ent, ref MapInitEvent args)
     {
-        //_actions.AddAction(ent, ref ent.Comp.ZLevelActionEntity, ent.Comp.ActionProto);
+        _actions.AddAction(ent, ref ent.Comp.ActionEntity, ent.Comp.ActionId);
         _meta.AddFlag(ent, MetaDataFlags.ExtraTransformEvents);
     }
 
     private void OnCompRemove(Entity<CEZLevelViewerComponent> ent, ref ComponentRemove args)
     {
-        //_actions.RemoveAction(ent.Comp.ZLevelActionEntity);
+        _actions.RemoveAction(ent.Comp.ActionEntity);
         _meta.RemoveFlag(ent, MetaDataFlags.ExtraTransformEvents);
 
         foreach (var eye in ent.Comp.Eyes)
@@ -121,13 +121,10 @@ public sealed partial class CEZLevelsSystem
             eyes.Add(newEye);
         }
 
-        // We constantly load the upper z-levels for the client so that you can quickly look up and climb stairs without PVS lag.
-        for (var i = 1; i <= MaxZLevelsAboveRendering; i++)
+        // We constantly load the upper z-level for the client so that you can quickly look up and climb stairs without PVS lag.
+        if (TryMapUp(map.Value, out var aboveMapUid))
         {
-            if (!TryMapOffset(map.Value, i, out var mapUidAbove))
-                break;
-
-            var newEye = SpawnAtPosition(_zEyeProto, new EntityCoordinates(mapUidAbove, globalPos));
+            var newEye = SpawnAtPosition(_zEyeProto, new EntityCoordinates(aboveMapUid, globalPos));
 
             Transform(newEye).GridTraversal = false;
             _viewSubscriber.AddViewSubscriber(newEye, actor.PlayerSession);
