+++ Content.Server/_CE/ZLevels/Mapping/CEZLevelMappingSystem.cs
@@ -44,7 +44,7 @@ public sealed class CEZLevelMappingSystem : EntitySystem
 
     private void OnMapInit(Entity<CEZLevelMapComponent> ent, ref MapInitEvent args)
     {
-        if (!_zLevels.TryGetZNetwork((ent, ent.Comp), out var network))
+        if (!_zLevels.TryGetZNetwork(ent, out var network))
             return;
 
         EntityManager.AddComponents(ent, network.Comp.Components);
