+++ Content.Server/_CE/ZLevels/PVS/CEPvsOverrideComponent.cs
@@ -3,9 +3,7 @@
  * https://github.com/space-wizards/space-station-14/blob/master/LICENSE.TXT
  */
 
-namespace Content.Server._CE.PVS;
+namespace Content.Server._CE.ZLevels.PVS;
 
 [RegisterComponent]
-public sealed partial class CEPvsOverrideComponent : Component
-{
-}
+public sealed partial class CEPvsOverrideComponent : Component;
