+++ Content.Shared/_CE/ZLevels/Core/Components/CEZLevelsNetworkComponent.cs
@@ -5,6 +5,7 @@
 
 using Content.Shared._CE.ZLevels.Core.EntitySystems;
 using Robust.Shared.GameStates;
+using Robust.Shared.Prototypes;
 
 namespace Content.Shared._CE.ZLevels.Core.Components;
 
@@ -17,6 +18,12 @@ public sealed partial class CEZLevelsNetworkComponent : Component
 {
     [ViewVariables, AutoNetworkedField]
     public readonly Dictionary<int, EntityUid?> ZLevels = new();
+
+    /// <summary>
+    /// Shared components for all zLevels maps
+    /// </summary>
+    [DataField(serverOnly: true)]
+    public ComponentRegistry Components = new();
 
     /// <remarks>
     /// Reversed version of <see cref="ZLevels"/>
