--- Content.Shared/_CE/ZLevels/Weapons/CEZLevelWeaponShootComponent.cs
@@ -1,6 +1,0 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared._CE.ZLevels.Weapons;
-
-[RegisterComponent, NetworkedComponent]
-public sealed partial class CEZLevelWeaponShootComponent : Component;
